requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
python-dotenv==1.0.1
selenium==4.26.1
pillow==11.0.0
